+++ galasa-simbank-tests/dev.galasa.simbank.manager/src/main/java/dev/galasa/simbank/manager/gherkin/Then.java
@@ -1,4 +1,4 @@
-package dev.galasa.simbank.manager.ghrekin;
+package dev.galasa.simbank.manager.gherkin;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
