+++ galasa-simbank-tests/dev.galasa.simbank.manager/src/main/java/dev/galasa/simbank/manager/gherkin/When.java
@@ -1,9 +1,13 @@
-package dev.galasa.simbank.manager.ghrekin;
+package dev.galasa.simbank.manager.gherkin;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Unique {
+public @interface When {
+
+    String regex();
+
+    String type();
     
 }
